+++ src/Generics/Deriving.hs
@@ -4,6 +4,7 @@ module Generics.Deriving (
     module Generics.Deriving.Base,
     module Generics.Deriving.Copoint,
     module Generics.Deriving.ConNames,
+    module Generics.Deriving.Default,
     module Generics.Deriving.Enum,
     module Generics.Deriving.Eq,
     module Generics.Deriving.Functor,
@@ -15,6 +16,7 @@ module Generics.Deriving (
 import Generics.Deriving.Base
 import Generics.Deriving.Copoint
 import Generics.Deriving.ConNames
+import Generics.Deriving.Default
 import Generics.Deriving.Enum
 import Generics.Deriving.Eq
 import Generics.Deriving.Functor